# pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Простой API для управления задачами",
        "title": "ToDo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/tasks": {
            "get": {
                "description": "Получить список всех задач",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить все задачи",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/taskRepository.Task"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка получения списка задач",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавить новую задачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создание задачи",
                "parameters": [
                    {
                        "description": "Новая задача",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TaskCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача создана",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ошибка добавления задачи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "description": "Обновить задачу по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Обновление задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённая задача",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.TaskCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача обновлена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка обновления задачи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить задачу по ID",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удаление задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача удалена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "неверный ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.TaskCreateRequest": {
            "description": "Структура запроса для создания или обновления задачи",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Описание задачи",
                    "type": "string"
                },
                "status": {
                    "description": "Статус задачи",
                    "type": "string"
                },
                "title": {
                    "description": "Заголовок задачи",
                    "type": "string"
                }
            }
        },
        "taskRepository.Task": {
            "description": "Структура ответа для возвращаемых данных о задаче",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Время создания задачи\n@example \"2023-04-01T12:00:00Z\"",
                    "type": "string"
                },
                "description": {
                    "description": "Описание задачи\n@example \"Пойти в магазин и купить хлеб\"",
                    "type": "string"
                },
                "id": {
                    "description": "ID задачи\n@example 1",
                    "type": "integer"
                },
                "status": {
                    "description": "Статус задачи\n@example \"new\"",
                    "type": "string"
                },
                "title": {
                    "description": "Заголовок задачи\n@example \"Купить хлеб\"",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Время последнего обновления задачи\n@example \"2023-04-02T14:00:00Z\"",
                    "type": "string"
                }
            }
        }
    }
}